organized all firebase work used in events into a controller. added a cleanup function to remove all old data.
